+++ EasySaveConsole/EasySave/App.xaml.cs
@@ -1,4 +1,5 @@
 ﻿using EasySave.Services;
+using EasySave.ViewModels;
 using System.Configuration;
 using System.Data;
 using System.Windows;
