+++ EasySaveConsole/EasySave/Helpers/ObservableObject.cs
@@ -1,0 +1,13 @@
+﻿using System.ComponentModel;
+
+namespace EasySave.Helpers;
+
+public class ObservableObject : INotifyPropertyChanged
+{
+    public event PropertyChangedEventHandler? PropertyChanged;
+
+    protected virtual void OnPropertyChanged(string propertyName)
+    {
+        PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
+    }
+}
