+++ EasySaveConsole/EasySave/Views/ConfigView.xaml.cs
@@ -1,5 +1,5 @@
 ﻿using EasySave.Models;
-using EasySave.Services;
+using EasySave.ViewModels;
 using System.Windows;
 using System.Windows.Controls;
